+++ modules/table/src/testFixtures/java/org/apache/ignite/internal/table/distributed/disaster/TestDisasterRecoveryUtils.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.table.distributed.disaster;
+
+import static org.apache.ignite.internal.lang.IgniteSystemProperties.enabledColocation;
+
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * Helper util class for disaster recovery tests.
+ */
+public class TestDisasterRecoveryUtils {
+
+    public static CompletableFuture<Void> resetPartitions(
+            DisasterRecoveryManager disasterRecoveryManager,
+            String zoneName,
+            String schemaName,
+            String tableName,
+            Set<Integer> partitionIds,
+            boolean manualUpdate,
+            long triggerRevision
+    ) {
+        if (enabledColocation()) {
+            return disasterRecoveryManager.resetPartitions(zoneName, partitionIds, manualUpdate, triggerRevision);
+        } else {
+            return disasterRecoveryManager
+                    .resetTablePartitions(zoneName, schemaName, tableName, partitionIds, manualUpdate, triggerRevision);
+        }
+    }
+}
